try to have a working workflow with OK tests
