+++ pyproject.toml
@@ -20,3 +20,4 @@ classifiers = [
 [project.urls]
 "Homepage" = "https://github.com/Ch1nkara/M1-2022-git-workflow-dubaret"
 "Bug Tracker" = "https://github.com/Ch1nkara/M1-2022-git-workflow-dubaret/issues"
+
